requirements.txt
-e src/
ipympl
ipywidgets
jupyterlab
matplotlib
pandas
pip
python-dateutil
pyyaml
scipy
